requirements.txt
flask==2.0.1
pandas==1.3.3
scikit-learn==0.24.2
python-dotenv==0.19.0
groq==0.4.2
pymongo==3.12.0